test(twenty-slack): keep single-consumer helpers local to their module
