--- packages/twenty-apps/public/slack/src/__tests__/types/app-runtime-mock-options.type.ts
@@ -1,7 +1,0 @@
-import { type AppConnection } from 'twenty-sdk/logic-function';
-
-export type AppRuntimeMockOptions = {
-  apiUrl: string;
-  workspaceId: string;
-  connections?: AppConnection[];
-};
