--- packages/twenty-apps/public/slack/src/__tests__/types/slack-api-mock-options.type.ts
@@ -1,5 +1,0 @@
-export type SlackApiMockOptions = {
-  botToken?: string;
-  botUserId?: string;
-  teamId?: string;
-};
