+++ packages/twenty-apps/public/slack/src/__tests__/utils/build-slack-route-payload.util.ts
@@ -1,7 +1,23 @@
+import { createHmac } from 'crypto';
+
 import { type RoutePayload } from 'twenty-sdk/define';
 
-import { signSlackRequest } from 'src/__tests__/utils/sign-slack-request.util';
+import { SLACK_TEST_WEBHOOK_SECRET } from 'src/__tests__/constants/slack-test-webhook-secret.constant';
+
 import { type SlackEventsRequestBody } from 'src/logic-functions/types/slack-events-request-body.type';
+
+const signSlackRequest = ({
+  rawBody,
+  timestampSeconds,
+  secret = SLACK_TEST_WEBHOOK_SECRET,
+}: {
+  rawBody: string;
+  timestampSeconds: number;
+  secret?: string;
+}): string =>
+  `v0=${createHmac('sha256', secret)
+    .update(`v0:${timestampSeconds}:${rawBody}`, 'utf8')
+    .digest('hex')}`;
 
 export const buildSlackRoutePayload = (
   body: SlackEventsRequestBody,
